move exceptions to their own package, ditch the ffi package.
