+++ java/lib/src/main/java/io/buttplug/MessageAttributes.java
@@ -1,4 +1,4 @@
-package io.buttplug.ffi;
+package io.buttplug;
 
 import io.buttplug.protos.ButtplugRsFfi;
 import io.buttplug.protos.ButtplugRsFfi.ServerMessage;
