+++ java/lib/src/main/java/io/buttplug/WebsocketConnectorOptions.java
@@ -1,4 +1,4 @@
-package io.buttplug.ffi;
+package io.buttplug;
 
 import java.net.URI;
