+++ java/lib/src/main/java/io/buttplug/exceptions/ButtplugConnectorException.java
@@ -1,4 +1,4 @@
-package io.buttplug;
+package io.buttplug.exceptions;
 
 public class ButtplugConnectorException extends ButtplugException {
     public ButtplugConnectorException(String msg) {
