+++ java/lib/src/main/java/io/buttplug/exceptions/ButtplugException.java
@@ -1,4 +1,4 @@
-package io.buttplug;
+package io.buttplug.exceptions;
 
 import io.buttplug.protos.ButtplugRsFfi.ServerMessage;
