+++ java/lib/src/main/java/io/buttplug/exceptions/ButtplugPingException.java
@@ -1,4 +1,4 @@
-package io.buttplug;
+package io.buttplug.exceptions;
 
 public class ButtplugPingException extends ButtplugException {
     public ButtplugPingException(String msg) {
